Use type alias syntax in test
