+++ pyproject.toml
@@ -180,6 +180,7 @@ dynamic_context = "test_function"
 omit = [
     "docs_src/response_model/tutorial003_04.py",
     "docs_src/response_model/tutorial003_04_py310.py",
+    "tests/_test_type_alias_depends.py",
 ]
 
 [tool.coverage.report]
